+++ internal/member/repository.go
@@ -40,7 +40,7 @@ func (m *MemberRepository) FindByEmail(ctx context.Context, db *gorm.DB, email s
 	return &member, nil
 }
 
-func (m *MemberRepository) FindByID(ctx context.Context, db *gorm.DB, ID uint32) (*model.Member, error) {
+func (m *MemberRepository) FindByID(ctx context.Context, db *gorm.DB, ID int64) (*model.Member, error) {
 	var member model.Member
 	err := db.WithContext(ctx).Where("id = ?", ID).First(&member).Error
 	if err != nil {
